requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
praw>=7.7.1
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
google-generativeai>=0.5.0
python-dotenv>=1.0.0
numpy>=1.26.0
tqdm>=4.66.1
pydantic>=2.5.0